pyproject.toml
[project]
name = "adk-branding-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10.0",
    "google-adk>=1.2.1",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.0",
    "litellm>=1.71.0",
    "numpy>=2.0.0",
    "openai>=1.50.0",
    "opencv-python>=4.10.0",
    "pillow>=11.2.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "replicate>=0.30.0",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "black>=24.4.0",
    "isort>=5.13.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

image = [
    "opencv-python>=4.10.0",  # Optional: Advanced computer vision (if needed later)
]

production = [
    "gunicorn>=22.0.0",
    "uvicorn>=0.30.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "config", "tools"]
include = [
    "main.py",
    "scripts/*",
    "data/*",
    "docs/*"
]